move to sorted blocks
