--- src/test/java/com/simplymadeapps/simple_logger_android/TestConstructors.java
@@ -1,35 +1,0 @@
-package com.simplymadeapps.simple_logger_android;
-
-import android.content.Context;
-
-import com.snatik.storage.Storage;
-
-import junit.framework.Assert;
-
-import org.junit.Test;
-
-import static org.powermock.api.mockito.PowerMockito.mock;
-
-/**
- * Created by stephenruda on 8/4/17.
- */
-
-public class TestConstructors {
-
-    @Test
-    public void testConstructors() {
-        // These are empty constructors with no assertations
-        TransferHelper th = new TransferHelper();
-    }
-
-    @Test
-    public void test_newStorageInstance() {
-        Assert.assertEquals(SimpleAmazonLogsHelper.newStorageInstance(mock(Context.class)).getClass(), Storage.class);
-    }
-
-    @Test
-    public void test_helperConstructor() {
-        SimpleAmazonLogsHelper helpers = new SimpleAmazonLogsHelper();
-        Assert.assertEquals(helpers.getClass(), SimpleAmazonLogsHelper.class);
-    }
-}
